requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
